refactor: 패스워드 확인 로직 간 문자열 하드코딩 변경
