+++ core/ui/src/main/res/values/style.xml
@@ -29,7 +29,7 @@
         <item name="navigationIcon">?android:attr/actionModeCloseDrawable</item>
         <item name="titleMargin">0dp</item>
         <item name="android:textSize">@dimen/text_large</item>
-        <item name="font">@font/nanumsquareneo_bold</item>
+        <item name="font">@font/nanumsquareneo_heavy</item>
     </style>
 
     <style name="cardView" parent="Widget.Material3.CardView.Elevated">
