+++ feature/manage/src/main/res/layout/fragment_manage_password.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.lst_1995.manage.ManagePasswordViewModel" />
+            type="com.lst_1995.manage.password.ManagePasswordViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -35,6 +35,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:inputType="text|textPassword"
+                android:maxLength="10"
                 android:text="@={viewModel.password}" />
         </com.google.android.material.textfield.TextInputLayout>
 
@@ -57,6 +58,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
                 android:inputType="text|textPassword"
+                android:maxLength="10"
                 android:text="@={viewModel.passwordCheck}" />
         </com.google.android.material.textfield.TextInputLayout>
 
@@ -90,14 +92,36 @@
             android:id="@+id/passwordMessageTv"
             style="@style/textView"
             android:layout_width="wrap_content"
-            android:layout_height="20dp"
+            android:layout_height="wrap_content"
             android:layout_marginStart="60dp"
             android:layout_marginEnd="60dp"
             android:layout_marginBottom="48dp"
+            android:textColor="?attr/colorError"
             app:layout_constraintBottom_toTopOf="@+id/checkBtn"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintHorizontal_bias="0.497"
             app:layout_constraintStart_toStartOf="parent"
             app:text="@{viewModel.passwordSetMessage}" />
+
+        <com.google.android.material.appbar.MaterialToolbar
+            style="@style/toolBar"
+            android:id="@+id/materialToolbar2"
+            android:layout_width="0dp"
+            android:layout_height="wrap_content"
+            app:title="@string/password_manage"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <ProgressBar
+            android:id="@+id/progressBar"
+            style="?android:attr/progressBarStyle"
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+            android:visibility="@{viewModel.loadingProgressVisibility}"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
